services/booking-service/requirements.txt
flask==3.0.0
kafka-python==2.0.2
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0